\name{dbt.Transforms-package}
\alias{dbt.Transforms-package}
\alias{dbt.Transforms-package}
\docType{package}
\title{
DataBionics Toolbox - Transforms
}
\description{
This package contains functions to transform data (like z-tranfsformation, ica, cdf etc.) 
}
\details{
\tabular{ll}{
Package: \tab dbt.Transforms\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-05-03\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}

cdftrans, mds, ica, slog.... 


}
\keyword{ package }