\name{relDiff2Yield}
\alias{relDiff2Yield}
\title{Relative differences to yields}
\description{
  This function calculates yields from relative differences.
}
\usage{
relDiff2Yield(relDiff)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{relDiff}{relative differences }
}

\value{
 returns the yield
}


\author{fm}



\seealso{rellDiffTrans}
