\name{relDiff4AB}
\alias{relDiff4AB}
\title{ RelDiff for two items }
\description{
  This function calculates the modification of relative differences referred to the middle basic value.
}
\usage{
relDiff4AB(a, b)
}
\arguments{
  \item{a}{ base item, A+B!=0.}
  \item{b}{ base item, A+B!=0.}
}

\value{
  returns the modification referred to the middle basic value.
}


\author{fm}

