\name{relDiff4LogRatio}
\alias{relDiff4LogRatio}

\title{ RelDiff for LogRatio}
\description{
  This function transforms LogRatio into RelDiff.
}
\usage{
relDiff4LogRatio(logRatio)
}

\arguments{
  \item{logRatio}{numeric matrix}
}

\value{
  returns the RelDiff of LogRatio
}

\author{fm}
