\name{relDiff4Percent}
\alias{relDiff4Percent}

\title{ RelDiff for percent}
\description{
  This function transforms percents in RelDiffs
}
\usage{
relDiff4Percent(percent)
}

\arguments{
  \item{percent}{a numeric matrix.}
}

\value{
 returns a matrix of the RelDiff.
}

\author{fm}
