\name{relDiff4Yield}
\alias{relDiff4Yield}

\title{ RelDiff for yield }
\description{
  This function transforms yields into relative differences.
}
\usage{
relDiff4Yield(yields)
}

\arguments{
  \item{yields}{a numeric matrix of yields, not in percent!}
}

\value{
  returns a matrix of relative differences
}

\author{ fm }
