\name{relDiffTrans}
\alias{relDiffTrans}

\title{RelDiff Transformation }
\description{
  This function transforms yields or percents in RelDiff. 
}
\usage{
relDiffTrans(yield)
}

\arguments{
  \item{yield}{a numeric matrix of yields}
}


\value{
  returns a matrix of RelDiff
}

\author{fm}
