\name{sloginvers}
\alias{sloginvers}
\title{inverse signed log}
\description{
 This function returns a inverse signed log 
}
\usage{
sloginvers(signedLog, m)
}
\arguments{
  \item{signedLog}{signed log}
  \item{m}{mantissa of log}
}

\value{
 returns the inverse signed log of data
}

\author{ fm }

