\name{sRelDiffTransInvers}
\alias{sRelDiffTransInvers}

\title{Inverse Signed RelDiff Transformation}
\description{
  This function is the inverse function to \code{sRelDiffTrans}.
}
\usage{
sRelDiffTransInvers(sRelDiff)
}

\arguments{
  \item{sRelDiff}{ a matrix of RelDiff.}
}


\value{
  returns a matrix of yields. 
}

\author{fm}
