\name{yield4LogRatio}
\alias{yield4LogRatio}

\title{ yields for LogRatio}
\description{
 This function transforms LogRatio into yields.
}
\usage{
yield4LogRatio(logRatio)
}

\arguments{
  \item{logRatio}{ a numeric matrix. }
}


\value{
  returns a matrix of yields.
}
\author{fm}
\note{ only approximately!
}
