\name{ztrans}
\alias{ztrans}

\title{Z-Transformation}
\description{
  This function returns z-transformed data. If data contains any NaN or Inf, \code{transNaN} will be called.
}
\usage{
ztrans(x)
}

\arguments{
  \item{x}{a vector or matrix of numeric data}
}

\value{
  returns a vector or matrix of z-transformed data.
}

\author{fm}

