\name{ztransNaN}
\alias{ztransNaN}

\title{Z-Transformation, ignoring NaNs}
\description{
   This function returns z-transformed data, NaN-values will be untransformed.
}
\usage{
ztransNaN(x)
}

\arguments{
  \item{x}{ a numeric dataarray  }
}

\value{
  returns a vector or matrix of z-transformed data  
}


\author{fm }
